gt - fixed tests, and added tests for unallowing different driver to access or edit this availability

+++ src/main/java/edu/ucsb/cs156/gauchoride/controllers/DriverAvailabilityController.java
@@ -84,9 +84,13 @@ public class DriverAvailabilityController extends ApiController{
                     required = true)  
                     @RequestParam Long id) 
     {
-        DriverAvailability availability;
-        availability = driverAvailabilityRepository.findById(id)
-                    .orElseThrow(() -> new EntityNotFoundException(DriverAvailability.class, id));
+        DriverAvailability availability = driverAvailabilityRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(DriverAvailability.class, id));
+
+        if (availability.getDriverId() != getCurrentUser().getUser().getId()){
+            throw new EntityNotFoundException(DriverAvailability.class, id);
+        }
+        
         return availability;
     }
 
@@ -100,12 +104,13 @@ public class DriverAvailabilityController extends ApiController{
                             @RequestParam Long id,
                             @RequestBody @Valid DriverAvailability incoming)
     {
-        DriverAvailability availability;
+        DriverAvailability availability = driverAvailabilityRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(DriverAvailability.class, id));
 
-        availability = driverAvailabilityRepository.findById(id)
-                    .orElseThrow(() -> new EntityNotFoundException(DriverAvailability.class, id));
+        if (availability.getDriverId() != getCurrentUser().getUser().getId()) {
+            throw new EntityNotFoundException(DriverAvailability.class, id);
+        }
 
-        availability.setDriverId(incoming.getDriverId());
         availability.setDay(incoming.getDay());
         availability.setStartTime(incoming.getStartTime());
         availability.setEndTime(incoming.getEndTime());
@@ -124,6 +129,10 @@ public class DriverAvailabilityController extends ApiController{
             @Parameter(name="id") @RequestParam Long id) {
         DriverAvailability availability = driverAvailabilityRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(DriverAvailability.class, id));
+
+        if (availability.getDriverId() != getCurrentUser().getUser().getId()) {
+            throw new EntityNotFoundException(DriverAvailability.class, id);
+        }
 
         driverAvailabilityRepository.delete(availability);
         return genericMessage("DriverAvailability with id %s deleted".formatted(id));
